finishing first phase
handled transfers errors

+++ src/Transactions/toBankTransfer.java
@@ -3,33 +3,43 @@ package Transactions;
 import Database.Account;
 import Database.bankAccount;
 import Database.database;
-import User.instaPayBankUser;
+import Database.walletAccount;
 import User.*;
 
-import static API.InstaPayAPI.getUser;
+import static API.DatabaseFunctions.getUser;
 
-public class toBankTransfer implements Transfer{
+public class toBankTransfer implements Transfer
+{
+    @Override
     public boolean transfer(Account acc, int amount, User currentuser)
     {
-        if(!database.bankDatabase.checkExistance(((bankAccount) acc).getBankAccountID()))
+        if(acc instanceof bankAccount)
         {
-            System.out.println("Couldn't find the user you want to transfer to!");
-            System.out.println();
-            return false;
-        }
-        else if(!database.instaDatabase.checkBalance(currentuser.getInstaID(),amount))
-        {
-            System.out.println("Your balance is not enough!");
-            System.out.println();
-            return false;
-        }
-        else {
-            database.bankDatabase.removeCredit(((instaPayBankUser) currentuser).getBankAccountID(),amount);
-            database.bankDatabase.addCredit(((bankAccount) acc).getBankAccountID(),amount);
+            if(!database.bankDatabase.checkExistance(((bankAccount) acc).getBankAccountID()))
+            {
+                System.out.println("Couldn't find the user you want to transfer to!");
+                System.out.println();
+                return false;
+            }
+            else if(!database.instaDatabase.checkBalance(currentuser.getInstaID(),amount))
+            {
+                System.out.println("Your balance is not enough!");
+                System.out.println();
+                return false;
+            }
             database.instaDatabase.removeCredit(currentuser.getInstaID(),amount);
+            database.bankDatabase.addCredit(((bankAccount) acc).getBankAccountID(),amount);
+            if(currentuser.getType().equals(userType.instaPayBankUser))
+            {
+                database.bankDatabase.removeCredit(((instaPayBankUser) currentuser).getBankAccountID(),amount);
+            }
+            else
+            {
+                Database.dummyWalletDatabase.removeCredit(currentuser.getMobileNumber(),amount);
+            }
             try
             {
-                database.instaDatabase.addCredit(getUser(acc.getUserName(),acc.getPassword()).getInstaID(), amount);
+                database.instaDatabase.addCredit(getUser(acc.getMobileNumber()).getInstaID(), amount);
             }
             catch(Exception e)
             {
@@ -37,8 +47,16 @@ public class toBankTransfer implements Transfer{
             }
             return true;
         }
-
-
+        else if(acc instanceof walletAccount)
+        {
+            return false;
+        }
+        else
+        {
+            System.out.println("Couldn't find the user you want to transfer to!");
+            System.out.println();
+            return false;
+        }
     }
 }
